+++ src/axios/instance.ts
@@ -1,7 +1,7 @@
 import axios from "axios";
 
-const instance = axios.create({
+const axiosInstance = axios.create({
   baseURL: process.env.NEXT_PUBLIC_API_URL,
 });
 
-export default instance;
+export default axiosInstance;
